move write_id
