+++ ydb/core/persqueue/common/partition_id.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <ydb/core/persqueue/write_id.h>
+#include <ydb/core/persqueue/public/write_id.h>
 #include <ydb/core/kafka_proxy/kafka_producer_instance_id.h> // TODO remove include - wrong dependency
 
 #include <util/generic/maybe.h>
