+++ ydb/core/persqueue/ya.make
@@ -5,7 +5,6 @@ SRCS(
     key.cpp
     write_meta.cpp
     dread_cache_service/caching_service.cpp
-    write_id.cpp
 )
 
 GENERATE_ENUM_SERIALIZATION(sourceid_info.h)
